Mongo index fix
Ignore the code, the schema index is useless. Was getting a memory overflow when searching > 700 pages. Created an index on (valid, deleted, datecreated) which are the query params in a basic pagination search. If there are ever 700+ pages while adding something else to that search query, it will have to be added to the index. Damn it I'm not a dba.