Use card name instead of _id to check uniqueness
This is needed due to alternate versions of the same climax. This will cause an issue if the user is using two different language variants of the same climax. There's no real way to account for this since english cards dont have jp name data, and jp cards might not have english name data.